html/browsers/browsing-the-web/navigating-across-documents/navigation-unload-data-url.html

<!doctype html>
<title>data: URL navigation started from unload handler must be ignored</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/helpers.js"></script>
<body>

<script>
promise_test(async () => {
  const iframe = await setupAndWaitForIframe();

  iframe.contentWindow.addEventListener("unload", () => {
    iframe.contentWindow.location.href = "data:text/html,unload<script>parent.postMessage('fail', '*');<\/script>";
  });

  iframe.src = `data:text/html,load<script>parent.postMessage('pass', '*')<\/script>`;
  assert_equals(await waitForMessage(iframe.contentWindow), "pass");
});
</script>
